feat!: `gix rev resolve --explain` (#427)
